Add `make helm-chart` target and use helm chart in `make run`
The Helm chart will auto-generate passwords, but includes all the dev certs
as defaults in values.yaml. Other default settings come from the role manifest
and not from the dev settings.

DOMAIN is a _required_ config variable.

UAA_ADMIN_CLIENT_SECRET must be supplied by the user and not be generated
(the same value needs to be passed to the CF chart).

The UAA port is marked as _public_ so that it becomes reachable from CF.