Fix mutual TLS certs when HA mode is true
When HA is set, we never updated the counts in helm; this means that the
secret generator didn't know about the correct counts, and therefore did
not generate certs which would apply to all the nodes in self-clustering
roles.  This fixes the issue by changing the secret generator to create
wildcard certificates (limited to the appropriate kubernetes headless
services for each role).  We also drop the sizing variables as they are
no longer necessary due to the above change.