Add pre-start script to wait for the database to accept TCP connections
This makes the UAA role not start until after the database accepts TCP
connections.  It makes UAA take slightly longer to start up, but avoids
crashing it when it fails to connect to the database.  The net effect is
that monit doesn't eventually give up and start backing off starting UAA
for ten minutes.