Remove KUBE_SIZING_MYSQL_COUNT environment variable
It is provided by fissile and always corresponds to the number of
mysql instances. This used to be used to set KUBE_MYSQL_CLUSTER_IPS
in configure-HA-hosts.sh. But this logic is only used for NATS
anymore, so there is no reason to restart all mysql-proxy pods
whenever the mysql replica count changes.

[jsc#CAP-1148]